+++ Mods/S7_Journal_6b1aee95-bd4c-4190-a919-4aa7096eaa2b/Story/RawFiles/Lua/BootstrapClient.lua
@@ -3,6 +3,7 @@
 --  =======
 
 Ext.Require("Auxiliary.lua")
+Ext.Require("TreasureTables.lua")
 
 --  ==================
 --  FETCH JOURNAL DATA
@@ -30,7 +31,7 @@ Ext.RegisterNetListener(IDENTIFIER, function (channel, payload)
     local journal = Ext.JsonParse(payload)
     if journal.ID == "CharacterOpenJournal" then
         FileName = journal.Data.fileName
-        S7DebugPrint("Dispatching BuildSpecs to UI-Components-Library.", "BootstrapClient")
+        S7Debug:Print("Dispatching BuildSpecs to UI-Components-Library.")
         local BuildSpecs = {["GMJournal"] = Rematerialize(journal.Data.content)}
         if not UCL.Journal.Exists then
             UCL.UCLBuild(BuildSpecs)
@@ -39,78 +40,3 @@ Ext.RegisterNetListener(IDENTIFIER, function (channel, payload)
         SaveJournal()
     end
 end)
-
---  ===============
---  TREASURE TABLES
---  ===============
-
-
-if not CENTRAL[IDENTIFIER].ModSettings.Uniques then
-    Ext.RegisterListener("StatsLoaded", function()
-        local treasureCat = {
-            ["Category"] = "I_BOOK_S7_JournalStat",
-            ["Items"] =
-            {
-                {
-                    ["ActPart"] = 5,
-                    ["MaxAmount"] = 1,
-                    ["MaxLevel"] = 0,
-                    ["MinAmount"] = 1,
-                    ["MinLevel"] = 5,
-                    ["Name"] = "BOOK_S7_JournalStat",
-                    ["Priority"] = 1,
-                    ["Unique"] = 0
-                },
-            }
-        }
-
-        Ext.UpdateTreasureCategory("I_BOOK_S7_JournalStat", treasureCat)
-
-        S7DebugPrint("Created Treasure Category for Journals", "BootstrapClient")
-
-        local targetTreasureTables = {
-            "ST_IngredientsTrader",
-            "ST_KitchenThings",
-            "ST_PaperWork"
-        }
-
-        for _, target in pairs(targetTreasureTables) do
-            local treasure = Ext.GetTreasureTable(target)
-            treasure.SubTables[#treasure.SubTables+1] = {
-                ["Categories"] = {
-                    {
-                        ["Common"] = 0,
-                        ["Divine"] = 0,
-                        ["Epic"] = 0,
-                        ["Frequency"] = 1,
-                        ["Legendary"] = 0,
-                        ["Rare"] = 0,
-                        ["TreasureCategory"] =  "I_BOOK_S7_JournalStat",
-                        ["Uncommon"] = 0,
-                        ["Unique"] = 0
-                    }
-                },
-                ["DropCounts"] = {
-                    {
-                        ["Amount"] = 1,
-                        ["Chance"] = 0
-                    },
-                    {
-                        ["Amount"] = 1,
-                        ["Chance"] = 1
-                    }
-                },
-                ["EndLevel"] = 0,
-                ["StartLevel"] = 0,
-                ["TotalCount"] = 1
-            }
-            Ext.UpdateTreasureTable(treasure)
-            S7DebugPrint("Added Journals to treasure-table: " .. target, "BootstrapClient")
-        end
-    end)
-else
-    Ext.RegisterListener("StatsLoaded", function()
-        local stat = Ext.GetStat("BOOK_S7_JournalStat")
-        stat.Unique = 1
-    end)
-end